refactor: item 도메인 상품 구매, 복구 행동 클래스 이름 변경 및 테스트 코드 수정
