+++ domain/item/src/main/java/store/nightmarket/domain/item/service/PurchaseItemDomainService.java
@@ -1,31 +1,51 @@
 package store.nightmarket.domain.item.service;
 
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 import store.nightmarket.common.domain.service.BaseDomainService;
 import store.nightmarket.domain.item.exception.ProductException;
 import store.nightmarket.domain.item.exception.QuantityException;
 import store.nightmarket.domain.item.model.ProductVariant;
-import store.nightmarket.domain.item.model.ShoppingBasketProduct;
+import store.nightmarket.domain.item.model.ShoppingBasket;
 import store.nightmarket.domain.item.service.dto.PurchaseItemDomainServiceDto.Event;
 import store.nightmarket.domain.item.service.dto.PurchaseItemDomainServiceDto.Input;
+import store.nightmarket.domain.item.valueobject.ProductVariantId;
 
 public class PurchaseItemDomainService
     implements BaseDomainService<Input, Event> {
 
     @Override
     public Event execute(Input input) {
-        ProductVariant productVariant = input.getProductVariant();
-        ShoppingBasketProduct shoppingBasketProduct = input.getShoppingBasketProduct();
+        List<ProductVariant> purchaseProductList = input.getPurchaseProductList();
+        ShoppingBasket shoppingBasket = input.getShoppingBasket();
 
-        if(productVariant.isNotSameAsProduct(shoppingBasketProduct)) {
-            throw new ProductException("Product variant is not same as shopping basket product");
+        if (purchaseProductList.isEmpty()) {
+            throw new ProductException("No product found");
         }
-        if (productVariant.isNotAbleToPurchase(shoppingBasketProduct)) {
-            throw new QuantityException("Product is not available to purchase");
-        }
-        productVariant.purchase(shoppingBasketProduct);
+        Map<ProductVariantId, ProductVariant> purchaseProductMap = purchaseProductList.stream()
+            .collect(Collectors.toMap(ProductVariant::getProductVariantId, Function.identity()));
+
+        shoppingBasket.getShoppingBasket()
+            .forEach(shoppingBasketProduct -> {
+                ProductVariantId productVariantId = shoppingBasketProduct.getVariantId();
+                if (!purchaseProductMap.containsKey(productVariantId)) {
+                    throw new ProductException(
+                        "Product variant is not same as shopping basket product: "
+                            + shoppingBasketProduct.getName());
+                }
+
+                ProductVariant purchaseProduct = purchaseProductMap.get(productVariantId);
+                if (purchaseProduct.isNotAbleToPurchase(shoppingBasketProduct)) {
+                    throw new QuantityException("Product is not available to purchase: "
+                        + shoppingBasketProduct.getName());
+                }
+                purchaseProduct.purchase(shoppingBasketProduct);
+            });
 
         return Event.builder()
-            .shoppingBasketProduct(shoppingBasketProduct)
+            .shoppingBasket(shoppingBasket)
             .build();
     }
