+++ domain/item/src/main/java/store/nightmarket/domain/item/service/dto/PurchaseItemDomainServiceDto.java
@@ -1,10 +1,10 @@
 package store.nightmarket.domain.item.service.dto;
 
+import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
 import store.nightmarket.domain.item.model.ProductVariant;
-import store.nightmarket.domain.item.model.ShoppingBasketProduct;
-import store.nightmarket.domain.item.valueobject.Quantity;
+import store.nightmarket.domain.item.model.ShoppingBasket;
 
 public class PurchaseItemDomainServiceDto {
 
@@ -12,16 +12,17 @@ public class PurchaseItemDomainServiceDto {
     @Builder
     public static class Input {
 
-        private final ProductVariant productVariant;
-        private final ShoppingBasketProduct shoppingBasketProduct;
+        List<ProductVariant> purchaseProductList;
+        ShoppingBasket shoppingBasket;
 
     }
+
 
     @Getter
     @Builder
     public static class Event {
 
-        private final ShoppingBasketProduct shoppingBasketProduct;
+        ShoppingBasket shoppingBasket;
 
     }
